+++ database/migrations/2024_11_08_180719_create_movimientos_table.php
@@ -16,11 +16,11 @@ return new class extends Migration
             $table->enum('tipo_movimiento', ['entrada','salida','transferencia']);
             $table->integer('cantidad_productos');
             $table->date('fecha_caducidad');
-            $table->integer('id_almacen_origen');
+            $table->unsignedBigInteger('id_almacen_origen');
             $table->foreign('id_almacen_origen')->references('id')->on('almacenes');
-            $table->integer('id_almacen_destino');
+            $table->unsignedBigInteger('id_almacen_destino');
             $table->foreign('id_almacen_destino')->references('id')->on('almacenes');
-            $table->integer('id_producto');
+            $table->unsignedBigInteger('id_producto');
             $table->foreign('id_producto')->references('id')->on('productos');
             $table->timestamps();
         });
